Fix script for clearing cache

+++ clearCache.sh
@@ -2,23 +2,20 @@
 
 set -e
 
-cd /container/home/www/zbk/cache/frbs
-
-if [ "$(pwd)" != "/container/home/www/zbk/cache/frbs" ]; then
-	echo "ERROR: not in the correct dir";
-else
-	echo "OK: correct dir $(pwd)"
-fi
+cd ./cache/donkeyCms
 
 for i in articleList categoriesForParent categoryTree css productDetail productExport productList
 do
-	echo "* changing to $i"
-	cd $i
-	echo "* pwd=$(pwd)"
-	echo "* removing all files"
-	rm -f *
-	echo "* removing done"
-	cd ../
+	if [ -d "$i" ]
+	then
+		echo "* changing to $i"
+		cd $i
+		echo "* pwd=$(pwd)"
+		echo "* removing all files"
+		rm -f *
+		echo "* removing done"
+		cd ../
+	fi
 done
